refactor(billing): extract shared invoice table and usage metric components
Move duplicated billing UI into reusable components and centralize status
label and badge styling for workspace billing and usage tabs.
